Create index/recipe pages and add all the content
Created the index page, where all of the links to the other recipe pages will be stored so that the index will work as the website's homepage.
All the respective recipe pages will link to the home page in case a user ever wants to go back and view other recipes. Every recipe page has a description, steps, and ingredients so that the user will have a understanding of what the food is, how to make it, and what the finished product will look like.  In retrospect I could have done one commit at A time to improve on my commit skills, but since I already have prior commit experience through a personal project I made, and since this project is simple enough, a single commit explaining everything makes sense to me.